V14: Integrations (Dynamics)
- Update property picker
- Update modal
- Update oauth configuration
