+++ src/Umbraco.Cms.Integrations.Crm.Dynamics/Client/src/modal/dynamics.modal-token.ts
@@ -4,7 +4,6 @@ import { DynamicsFormPickerConfiguration } from "../types/types";
 
 export type DynamicsFormPickerModalData = {
     headline: string;
-    config: DynamicsFormPickerConfiguration | undefined;
 }
 
 export type DynamicsFormPickerModalValue = {
